refactor: 記事の slug/permalink 導出を共通化し不整合を解消
- listAllRoutes の glob を .md/.mdx 両対応にし .md 記事の取りこぼしを修正
- RSS のリンクを article.permalink に統一しカスタム permalink の不一致を解消
- Frontmatter 型を types.ts に集約し存在しない型 import を修正
- buildPermalink ヘルパーを新設し permalink 生成の重複を排除
- formatDate の無効日付判定を明示化
- テストの配置ずれ・重複を解消し buildPermalink/無効日付のテストを追加
