+++ src/lib/content/articles.ts
@@ -28,11 +28,12 @@ export async function getAllPublishedArticles(): Promise<Article[]> {
 
 /**
  * 指定された記事に関連する記事のリストを取得します。
- * 公開日時の近い記事が優先されます。
+ * 関連度はタグやカテゴリではなく「公開日時の近さ」で判定し、
+ * 対象記事に近い日付の記事から順に返します。
  * @param target 基準となる記事オブジェクト。
  * @param allArticles すべての記事の配列。
  * @param max 取得する関連記事の最大数。デフォルトは5です。
- * @returns 関連記事の配列。
+ * @returns 公開日が近い順に並んだ関連記事の配列。
  */
 export function getRelatedArticles(
   target: Article,
