+++ src/lib/content/types.ts
@@ -16,3 +16,17 @@ export type Heading = {
 export type Article = CollectionEntry<'articles'> & {
   permalink: string;
 };
+
+/**
+ * 記事フロントマターの型定義。
+ * Astro の Content Collection を経由せず、生の Markdown/MDX を
+ * gray-matter や import.meta.glob で読み込む箇所で使用します。
+ * content.config.ts の articles スキーマに対応します。
+ */
+export type Frontmatter = {
+  title?: string;
+  published_at?: Date;
+  summary?: string;
+  permalink?: string;
+  draft?: boolean;
+};
